[FAB2023-1242] fix: 20240924_13 - 거버넌스 용어사전 삭제 얼럿 수정

+++ project/ovp/frontend/ovp/components/govern/glossary/glossary-info.vue
@@ -240,8 +240,16 @@ function syncEditDataWithGlossary() {
 }
 
 async function removeGlossary() {
-  await deleteGlossary(glossary.id);
-  await getGlossaries();
+  if (confirm("데이터모델을 삭제 하시겠습니까?")) {
+    await deleteGlossary(glossary.id);
+    await getGlossaries()
+      .then(() => {
+        alert("삭제되었습니다.");
+      })
+      .catch((error) => {
+        console.error("삭제 중 오류 발생: ", error);
+      });
+  }
 }
 
 async function updateGlossary(
